Main

Documentation

data Node

A Node is comprised of its index, the next index and a Bool showing it was reconnected or not. Any algorithm relaying on this data type should have access to a Map or List where it can fetch Node references from both index and nextIndex

type NodeSystem = Map Int Node

Just wraps a Map Int Node

makeNode

Arguments

:: Int
Node index

-> Int
Node index this node points to

-> (Int, Node)
As we are creating our node and will stuff them into a Map Int Node, here is the association

Helper just to not instantiate a Node by hand.

buildSystem :: [Int] -> NodeSystem

Given a list of Int, transform into a NodeSystem

reconnectIfBad :: NodeSystem -> Node -> NodeSystem

This function simply produces a new NodeSystem with the given Node updated with reconnected = True and nextIndex = 0, given that the default behavior for deep Bad nodes is just point them straight to a already known Good Node (Node 0 _ _).

findLastBad :: NodeSystem -> Node -> Maybe Node

Recursive function that given a pool of nodes (a NodeSystem) - possibly all nodes - will walk over the possible path through nextIndex.

The intention here is to return the deepest bad node in starting from a given Node.

Returns Nothing if the deepest Node is actually a Good Node.

solve :: NodeSystem -> Int

This is the code function that simply foldl over a NodeSystem producing a new one with all Good Nodes with minimum moves. Then it how many nodes where reconnected.

> (solve $ buildSystem [1,2,1] ) == 1

main :: IO ()